A better Sync Model (Part 1 of 3)
Starting to working in the better sync model.
A seperate class to check against time sync locally and remote.
Part 1 of 3 changes to come.
These changes work soley on the Subscriber classes and are  not complete. The subscriber tests will be the working model for the otehr areas of the module